mini-rag-baseline-stage2/requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
datasets
tqdm
numpy
evaluate
